+++ src/Professores.java
@@ -1,8 +1,6 @@
-import java.time.LocalDate;
-
 public class Professores extends Usuarios {
-    private String cargoAcademico;
-    private String cargoMinistrado;
+    private final String cargoAcademico;
+    private final String cargoMinistrado;
 
     public Professores(String nomeCompleto, String email, String telefone, String senha, String matricula,
             String cargoAcademico, String cargoMinistrado) {
@@ -17,17 +15,6 @@ public class Professores extends Usuarios {
 
     public String getCargoMinistrado() {
         return cargoMinistrado;
-    }
-
-    public int getAnoDeIngresso() {
-        String ano = matricula.substring(0, 2);
-        int anoAtual = LocalDate.now().getYear() % 100;
-        if (Integer.parseInt(ano) > anoAtual) {
-            ano = "19" + ano;
-        } else {
-            ano = "20" + ano;
-        }
-        return Integer.parseInt(ano);
     }
 
     @Override
