+++ src/Servidores.java
@@ -1,9 +1,7 @@
-import java.time.LocalDate;
-
 public class Servidores extends Usuarios {
 
-    private String cargoExercido;
-    private String departamento;
+    private final String cargoExercido;
+    private final String departamento;
 
     public Servidores(String nomeCompleto, String email, String telefone, String senha, String matricula,String cargoExercido,String departamento) {
         super(nomeCompleto, email, telefone, senha, matricula);
@@ -19,16 +17,6 @@ public class Servidores extends Usuarios {
         return departamento;
     }
 
-    public int getAnoDeIngresso() {
-        String ano = matricula.substring(0, 2);
-        int anoAtual = LocalDate.now().getYear() % 100;
-        if (Integer.parseInt(ano) > anoAtual) {
-            ano = "19" + ano;
-        } else {
-            ano = "20" + ano;
-        }
-        return Integer.parseInt(ano);
-    }
 
     public String toString() {
         String informacoes = "";
@@ -37,8 +25,8 @@ public class Servidores extends Usuarios {
         informacoes += "EMAIL: " + email + "\n";
         informacoes += "TELEFONE: " + telefone + "\n";
         informacoes += "ANO DE INGRESSO: " + getAnoDeIngresso() + "\n";
-        informacoes += "CARGO EXERCIDO: " + getCargoExercido() + "\n";
-        informacoes += "DEPARTAMENTO: " + getDepartamento() + "\n";
+        informacoes += "CARGO EXERCIDO: " + cargoExercido + "\n";
+        informacoes += "DEPARTAMENTO: " + departamento + "\n";
         return informacoes;
     }
